requirements.txt
Django==2.0.2
django-extensions==1.9.9
django-filter==1.1.0
django-rest-auth==0.9.3
djangorestframework==3.7.7
Markdown==2.6.11
Pillow==5.0.0
pytz==2018.3
six==1.11.0
typing==3.6.4
Werkzeug==0.14.1
